Rename: 기존 historyApi가 아닌 각 역할별 Api에서 History 기능 추가, 그에 따른 페이지, 모달 수정
